Change toLong

+++ facia-tool/app/Global.scala
@@ -26,7 +26,7 @@ object Global extends FaciaToolLifecycle with GlobalSettings with CloudWatchAppl
     ("solr-content-api-calls", ContentApiMetrics.HttpTimingMetric.count.toDouble),
     ("elastic-content-api-timeouts", ContentApiMetrics.ElasticHttpTimeoutCountMetric.getAndReset.toDouble),
     ("solr-content-api-timeouts", ContentApiMetrics.HttpTimeoutCountMetric.getAndReset.toDouble),
-    ("content-api-404", ContentApiMetrics.ContentApi404Metric.getAndReset.toLong)
+    ("content-api-404", ContentApiMetrics.ContentApi404Metric.getAndReset.toDouble)
   )
 
   override def onLoadConfig(config: Configuration, path: File, classloader: ClassLoader, mode: Mode.Mode): Configuration = {
